requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
black==22.6.0
click==8.1.3
httptools==0.4.0
iso8601==1.0.2
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.2
pypika-tortoise==0.1.6
pytz==2022.2.1
sanic==22.6.2
sanic-routing==22.3.0
tomli==2.0.1
tortoise-orm==0.19.2
typing-extensions==4.3.0
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
